+++ selenium/test_signup.py
@@ -1,18 +1,18 @@
-import time
 from pathlib import Path
 
 from helpers import createDriver
-from helpers import signup
+from helpers import delete_account
 from helpers import printNotice
 from helpers import printSuccess
-from helpers import delete_account
+from helpers import signup
 
 FRONTEND_URL = "http://localhost:5173"
 
 TEST_EMAIL = "rat8@njit.edu"
-TEST_PASSWORD = "password"   
+TEST_PASSWORD = "password"
 
-def test_signup(driver=None):
+
+def test_signup(driver=None, keep_driver=False):
     """Test signup here"""
 
     # ---------- Create driver for standalone tests ----------
@@ -27,11 +27,11 @@ def test_signup(driver=None):
     try:
         signup(driver)
         printSuccess(f"No errors in {Path(__file__).name}: {driver.current_url} \n")
-        
-    
+        return driver
+
     # ----------- Quit driver for standalone tests -----------
     finally:
-        if is_standalone_test:
+        if is_standalone_test and not keep_driver:
             delete_account(driver)
             driver.quit()
             print("Browser closed \n")
